payment перенесен из school в users
