+++ school/urls.py
@@ -15,10 +15,4 @@ urlpatterns = [
     path("lessons/<int:pk>/", views.LessonDetailAPIView.as_view(), name="lessons_detail"),
     path("lessons/update/<int:pk>/", views.LessonUpdateAPIView.as_view(), name="lessons_update"),
     path("lessons/delete/<int:pk>/", views.LessonDeleteAPIView.as_view(), name="lessons_delete"),
-
-    path("payment/", views.PaymentListAPIView.as_view(), name="payment"),
-    path("payment/create/", views.PaymentCreateAPIView.as_view(), name="payment_create"),
-    path("payment/<int:pk>/", views.PaymentDetailAPIView.as_view(), name="payment_detail"),
-    path("payment/update/<int:pk>/", views.PaymentUpdateAPIView.as_view(), name="payment_update"),
-    path("payment/delete/<int:pk>/", views.PaymentDeleteAPIView.as_view(), name="payment_delete"),
 ] + router.urls
