+++ school/views.py
@@ -1,7 +1,6 @@
 from rest_framework import viewsets, generics
-from school.models import Course, Lesson, Payment
-from school.serializers import CourseSerializer, LessonSerializer, PaymentSerializer
-from rest_framework.filters import SearchFilter, OrderingFilter
+from school.models import Course, Lesson
+from school.serializers import CourseSerializer, LessonSerializer
 
 
 class CourseViewSet(viewsets.ModelViewSet):
@@ -37,34 +36,3 @@ class LessonListAPIView(generics.ListAPIView):
 class LessonDetailAPIView(generics.RetrieveAPIView):
     serializer_class = LessonSerializer
     queryset = Lesson.objects.all()
-
-
-# Payment
-class PaymentCreateAPIView(generics.CreateAPIView):
-    serializer_class = PaymentSerializer
-
-
-# PATCH
-class PaymentUpdateAPIView(generics.UpdateAPIView):
-    serializer_class = PaymentSerializer
-    queryset = Payment.objects.all()
-
-
-# DELETE
-class PaymentDeleteAPIView(generics.DestroyAPIView):
-    serializer_class = PaymentSerializer
-    queryset = Payment.objects.all()
-
-
-class PaymentListAPIView(generics.ListAPIView):
-    serializer_class = PaymentSerializer
-    queryset = Payment.objects.all()
-
-    filter_backends = [SearchFilter, OrderingFilter]
-    ordering_fields = ['payment_date']
-    search_fields = ['paid_course', 'paid_lesson', 'payment_method']
-
-
-class PaymentDetailAPIView(generics.RetrieveAPIView):
-    serializer_class = PaymentSerializer
-    queryset = Payment.objects.all()
