+++ users/management/commands/add_payment.py
@@ -1,5 +1,6 @@
 from django.core.management.base import BaseCommand
-from school.models import Payment, Course
+from school.models import Course
+from users.models import Payment
 from django.core.management import call_command
