+++ users/serializers.py
@@ -1,6 +1,11 @@
 from rest_framework import serializers
-from .models import CustomUser
-from school.serializers import PaymentSerializer
+from .models import CustomUser, Payment
+
+
+class PaymentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Payment
+        fields = "__all__"
 
 
 class CustomUserSerializer(serializers.ModelSerializer):
